Added a check to see if a row is null/commented before turning it into a SQL statement. It's not a full solution yet because it basically just checks if it's null or a string and if it is it is ignored - so before a string PK was valid and now it isn't.
Also had to add the Excel Data Reader nuget package to the WFP project even though it doesn't use it - the references were super screwed up in the solution and this was the only way I could get it to work. Seems like it's a bug in .NET Standard libraries.